added minidashboard and styling
