+++ client/src/components/HowPrepared/HowPrepared.css
@@ -25,3 +25,8 @@
   right: 0px;
   bottom: 0px;
 }
+
+@media (max-width: 370px) {
+  .how-prepared-rating-selector {
+  }
+}
